refactor(lt_random_bytes, lt_port_random_bytes): change type of count to size_t

+++ hal/port/stm32/lt_port_stm32_nucleo_l432kc.c
@@ -23,13 +23,13 @@ RNG_HandleTypeDef rng;
 // SPI handle declaration
 SPI_HandleTypeDef SpiHandle;
 
-lt_ret_t lt_port_random_bytes(void *buff, uint16_t count)
+lt_ret_t lt_port_random_bytes(void *buff, size_t count)
 {
-    uint16_t bytes_left = count;
+    size_t bytes_left = count;
     uint8_t *buff_ptr = buff;
     while (bytes_left) {
         uint32_t random_data = HAL_RNG_GetRandomNumber(&rng);
-        uint16_t cpy_cnt = bytes_left < sizeof(random_data) ? bytes_left : sizeof(random_data);
+        size_t cpy_cnt = bytes_left < sizeof(random_data) ? bytes_left : sizeof(random_data);
         memcpy(buff_ptr, &random_data, cpy_cnt);
         bytes_left -= cpy_cnt;
         buff_ptr += cpy_cnt;
