+++ include/libtropic_port.h
@@ -9,6 +9,8 @@
 #ifndef LT_LIBTROPIC_PORT_H
 #define LT_LIBTROPIC_PORT_H
 
+#include <stdlib.h>
+
 #include "libtropic_common.h"
 
 /**
@@ -95,7 +97,7 @@ lt_ret_t lt_port_delay_on_int(lt_l2_state_t *s2, uint32_t ms);
  * @param len         Number of random bytes
  * @return lt_ret_t   LT_OK if success, otherwise returns other error code.
  */
-lt_ret_t lt_port_random_bytes(void *buff, uint16_t count);
+lt_ret_t lt_port_random_bytes(void *buff, size_t count);
 
 /** @} */  // end of group_port_functions
